+++ turses/__init__.py
@@ -4,4 +4,4 @@
 #         Licensed under the GPL License. See LICENSE for full details.       #
 ###############################################################################
 
-__version__ = "0.0.7"
+__version__ = "0.0.8"